Build two versions RandomX
Build two versions of the randomx NIFs one for the 512MIB dataset
(rx512) and one for the 4096MiB dataset (rx4096)
