+++ apps/arweave/c_src/Makefile
@@ -18,9 +18,6 @@ ERTS_INCLUDE_DIR ?= $(shell erl -noshell -eval 'io:format("~ts/erts-~ts/include/
 ERL_INTERFACE_INCLUDE_DIR ?= $(shell erl -noshell -eval 'io:format("~ts", [code:lib_dir(erl_interface, include)]).' -s init stop)
 ERL_INTERFACE_LIB_DIR ?= $(shell erl -noshell -eval 'io:format("~ts", [code:lib_dir(erl_interface, lib)]).' -s init stop)
 
-C_SRC_DIR = $(CURDIR)
-C_SRC_OUTPUT ?= $(CURDIR)/../priv/arweave.so
-
 # System type and C compiler/flags.
 
 UNAME_SYS := $(shell uname -s)
@@ -58,10 +55,26 @@ ifneq (, $(shell which pkg-config))
 	LDFLAGS += `pkg-config --libs openssl`
 endif
 
-CFLAGS += -fPIC -I $(ERTS_INCLUDE_DIR) -I $(ERL_INTERFACE_INCLUDE_DIR) -I /usr/local/include -I ../lib/RandomX/src
-CXXFLAGS += -fPIC -I $(ERTS_INCLUDE_DIR) -I $(ERL_INTERFACE_INCLUDE_DIR) -I ../lib/RandomX/src -std=c++11
+C_SRC_DIR = $(CURDIR)
 
-LDLIBS += -L $(ERL_INTERFACE_LIB_DIR) -L /usr/local/lib -lei -lssl -lcrypto ../lib/RandomX/build/librandomx.a
+CFLAGS += -fPIC -I $(ERTS_INCLUDE_DIR) -I $(ERL_INTERFACE_INCLUDE_DIR) -I /usr/local/include -I ../lib/RandomX/src -I $(C_SRC_DIR)
+CXXFLAGS += -fPIC -I $(ERTS_INCLUDE_DIR) -I $(ERL_INTERFACE_INCLUDE_DIR) -I ../lib/RandomX/src -std=c++11
+LDLIBS += -L $(ERL_INTERFACE_LIB_DIR) -L /usr/local/lib -lei -lssl -lcrypto 
+
+
+RX512_OUTPUT ?= $(CURDIR)/../priv/rx512_arweave.so
+RX4096_OUTPUT ?= $(CURDIR)/../priv/rx4096_arweave.so
+VDF_OUTPUT ?= $(CURDIR)/../priv/vdf_arweave.so
+
+COMMON_RANDOMX_SOURCES = $(wildcard $(C_SRC_DIR)/randomx/*.c $(C_SRC_DIR)/randomx/*.cpp)
+RX512_SOURCES = $(COMMON_RANDOMX_SOURCES) $(wildcard $(C_SRC_DIR)/*.c $(C_SRC_DIR)/randomx/rx512/*.c)
+RX4096_SOURCES = $(COMMON_RANDOMX_SOURCES) $(wildcard $(C_SRC_DIR)/*.c $(C_SRC_DIR)/randomx/rx4096/*.c)
+VDF_SOURCES = $(wildcard $(C_SRC_DIR)/*.c $(C_SRC_DIR)/vdf/*.c $(C_SRC_DIR)/vdf/*.cpp)
+
+RX512_OBJECTS = $(addsuffix .o, $(basename $(RX512_SOURCES)))
+RX4096_OBJECTS = $(addsuffix .o, $(basename $(RX4096_SOURCES)))
+VDF_OBJECTS = $(addsuffix .o, $(basename $(VDF_SOURCES)))
+
 
 # Verbosity.
 
@@ -74,15 +87,20 @@ cpp_verbose = $(cpp_verbose_$(V))
 link_verbose_0 = @echo " LD    " $(@F);
 link_verbose = $(link_verbose_$(V))
 
-ALL_LIB_SOURCES := $(wildcard $(C_SRC_DIR)/*.c $(C_SRC_DIR)/*.cpp)
-ALL_LIB_OBJECTS = $(addsuffix .o, $(basename $(ALL_LIB_SOURCES)))
-
 COMPILE_C = $(c_verbose) $(CC) $(CFLAGS) $(CPPFLAGS) -c
 COMPILE_CPP = $(cpp_verbose) $(CXX) $(CXXFLAGS) $(CPPFLAGS) -c
 
-$(C_SRC_OUTPUT): $(ALL_LIB_OBJECTS)
+$(RX512_OUTPUT): $(RX512_OBJECTS)
 	@mkdir -p $(BASEDIR)/priv/
-	$(link_verbose) $(CXX) $(ALL_LIB_OBJECTS) $(LDFLAGS) $(LDLIBS) -shared -o $(C_SRC_OUTPUT)
+	$(link_verbose) $(CXX) $(RX512_OBJECTS) $(LDFLAGS) $(LDLIBS) ../lib/RandomX/build512/librandomx512.a -shared -o $(RX512_OUTPUT)
+
+$(RX4096_OUTPUT): $(RX4096_OBJECTS)
+	@mkdir -p $(BASEDIR)/priv/
+	$(link_verbose) $(CXX) $(RX4096_OBJECTS) $(LDFLAGS) $(LDLIBS) ../lib/RandomX/build4096/librandomx4096.a -shared -o $(RX4096_OUTPUT)
+
+$(VDF_OUTPUT): $(VDF_OBJECTS)
+	@mkdir -p $(BASEDIR)/priv/
+	$(link_verbose) $(CXX) $(VDF_OBJECTS) $(LDFLAGS) $(LDLIBS) -shared -o $(VDF_OUTPUT)
 
 %.o: %.c
 	$(COMPILE_C) $(OUTPUT_OPTION) $<
@@ -96,8 +114,10 @@ $(C_SRC_OUTPUT): $(ALL_LIB_OBJECTS)
 %.o: %.cpp
 	$(COMPILE_CPP) $(OUTPUT_OPTION) $<
 
+all: $(RX512_OUTPUT) $(RX4096_OUTPUT) $(VDF_OUTPUT)
+
 clean:
-	@rm -f $(C_SRC_OUTPUT) $(ALL_LIB_OBJECTS)
+	@rm -f $(RX512_OUTPUT) $(RX4096_OUTPUT) $(VDF_OUTPUT) $(RX512_OBJECTS) $(RX4096_OBJECTS) $(VDF_OBJECTS)
