+++ apps/arweave/c_src/randomx/randomx_long_with_entropy.h
@@ -1,3 +1,6 @@
+#ifndef RANDOMX_LONG_WITH_ENTROPY_H
+#define RANDOMX_LONG_WITH_ENTROPY_H
+
 #include "randomx.h"
 
 #define RANDOMX_ENTROPY_SIZE (256*1024)
@@ -12,3 +15,5 @@ RANDOMX_EXPORT void randomx_decrypt_chunk(randomx_vm *machine, const unsigned ch
 #if defined(__cplusplus)
 }
 #endif
+
+#endif // RANDOMX_LONG_WITH_ENTROPY_H
