+++ apps/arweave/c_src/vdf/ar_vdf_nif.c
@@ -1,3 +1,9 @@
+#include <erl_nif.h>
+#include <string.h>
+#include <openssl/sha.h>
+#include <ar_nif.h>
+#include "vdf.h"
+
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 //    SHA
 ////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -57,7 +63,6 @@ static ERL_NIF_TERM vdf_parallel_sha_verify_with_reset_nif(
 		return enif_make_badarg(envPtr);
 	}
 
-	// copypasted from vdf_sha2_nif
 	if (!enif_inspect_binary(envPtr, argv[0], &Salt)) {
 		return enif_make_badarg(envPtr);
 	}
@@ -113,12 +118,24 @@ static ERL_NIF_TERM vdf_parallel_sha_verify_with_reset_nif(
 	// NOTE last paramemter will be array later
 	size_t outCheckpointSize = VDF_SHA_HASH_SIZE*(1+checkpointCount)*(1+skipCheckpointCount);
 	ERL_NIF_TERM outputTermCheckpoint;
-	unsigned char* outCheckpoint = enif_make_new_binary(envPtr, outCheckpointSize, &outputTermCheckpoint);
-	bool res = vdf_parallel_sha_verify_with_reset(Salt.data, Seed.data, checkpointCount, skipCheckpointCount, hashingIterations, InRes.data, InCheckpoint.data, outCheckpoint, ResetSalt.data, ResetSeed.data, maxThreadCount);
+	unsigned char* outCheckpoint = enif_make_new_binary(
+		envPtr, outCheckpointSize, &outputTermCheckpoint);
+	bool res = vdf_parallel_sha_verify_with_reset(
+		Salt.data, Seed.data, checkpointCount, skipCheckpointCount, hashingIterations,
+		InRes.data, InCheckpoint.data, outCheckpoint, ResetSalt.data, ResetSeed.data,
+		maxThreadCount);
 	// TODO return all checkpoints
 	if (!res) {
-		return error(envPtr, "verification failed");
+		return error_tuple(envPtr, "verification failed");
 	}
 
 	return ok_tuple(envPtr, outputTermCheckpoint);
 }
+
+static ErlNifFunc nif_funcs[] = {
+	{"vdf_sha2_nif", 5, vdf_sha2_nif, ERL_NIF_DIRTY_JOB_CPU_BOUND},
+	{"vdf_parallel_sha_verify_with_reset_nif", 10, vdf_parallel_sha_verify_with_reset_nif,
+		ERL_NIF_DIRTY_JOB_CPU_BOUND}
+};
+
+ERL_NIF_INIT(ar_vdf_nif, nif_funcs, NULL, NULL, NULL, NULL);
