+++ rebar.config
@@ -50,7 +50,6 @@
 		{copy, "bin/logs", "bin/logs"},
 		{copy, "bin/debug-logs", "bin/debug-logs"},
 		{copy, "bin/check-nofile", "bin/check-nofile"},
-		{copy, "apps/arweave/lib/RandomX/build/randomx-benchmark", "bin/randomx-benchmark"},
 		{copy, "data/not_found.html", "data/not_found.html"},
 		{copy, "data/hash_list_1_0", "data/hash_list_1_0"},
 		{copy, "data/genesis_wallets.csv", "data/genesis_wallets.csv"},
@@ -377,18 +376,33 @@
 ]}.
 
 {pre_hooks, [
+	% Build for randomx512 configuration
 	{"(darwin|linux|freebsd|netbsd|openbsd)", compile,
-		"bash -c \"mkdir -p apps/arweave/lib/RandomX/build && cd apps/arweave/lib/RandomX/build && cmake -DRANDOMX_ARGON_MEMORY=262144 -DRANDOMX_DATASET_BASE_SIZE=536870912 .. > /dev/null\""},
-	{"(darwin)", compile, "make randomx -C apps/arweave/lib/RandomX/build"},
-	{"(linux)", compile, "make -C apps/arweave/lib/RandomX/build"},
-	{"(freebsd|netbsd|openbsd)", compile, "gmake -C apps/arweave/lib/RandomX/build"},
-	{"(linux)", compile, "env AR=gcc-ar make -C apps/arweave/c_src"},
-	{"(darwin)", compile, "make -C apps/arweave/c_src"},
-	{"(freebsd|netbsd|openbsd)", compile, "gmake -C apps/arweave/c_src"}
+		"bash -c \"mkdir -p apps/arweave/lib/RandomX/build512 && cd apps/arweave/lib/RandomX/build512 && cmake -DRANDOMX_ARGON_MEMORY=262144 -DRANDOMX_DATASET_BASE_SIZE=536870912 .. > /dev/null\""},
+	{"(darwin)", compile, "make randomx -C apps/arweave/lib/RandomX/build512"},
+	{"(linux)", compile, "make -C apps/arweave/lib/RandomX/build512"},
+	{"(freebsd|netbsd|openbsd)", compile, "gmake -C apps/arweave/lib/RandomX/build512"},
+	{"(darwin|linux|freebsd|netbsd|openbsd)", compile, "bash -c \"cd apps/arweave/lib/RandomX/build512 && mv librandomx.a librandomx512.a\""},
+	% Build for randomx4096 configuration
+	{"(darwin|linux|freebsd|netbsd|openbsd)", compile,
+		"bash -c \"mkdir -p apps/arweave/lib/RandomX/build4096 && cd apps/arweave/lib/RandomX/build4096 && cmake -DRANDOMX_ARGON_MEMORY=524288 -DRANDOMX_DATASET_BASE_SIZE=4294967296 .. > /dev/null\""},
+	{"(darwin)", compile, "make randomx -C apps/arweave/lib/RandomX/build4096"},
+	{"(linux)", compile, "make -C apps/arweave/lib/RandomX/build4096"},
+	{"(freebsd|netbsd|openbsd)", compile, "gmake -C apps/arweave/lib/RandomX/build4096"},
+	{"(darwin|linux|freebsd|netbsd|openbsd)", compile, "bash -c \"cd apps/arweave/lib/RandomX/build4096 && mv librandomx.a librandomx4096.a\""},
+	% Compile NIFs
+	{"(linux)", compile, "env AR=gcc-ar make all -C apps/arweave/c_src"},
+	{"(darwin)", compile, "make all -C apps/arweave/c_src"},
+	{"(freebsd|netbsd|openbsd)", compile, "gmake all -C apps/arweave/c_src"}
 ]}.
 {post_hooks, [
-	{"(linux|darwin)", clean, "bash -c \"if [ -d apps/arweave/lib/RandomX/build ]; then make -C apps/arweave/lib/RandomX/build clean; fi\""},
-	{"(freebsd|netbsd|openbsd)", clean, "bash -c \"if [ -d apps/arweave/lib/RandomX/build ]; then gmake -C apps/arweave/lib/RandomX/build clean; fi\""},
+	% Clean randomx512
+	{"(linux|darwin)", clean, "bash -c \"if [ -d apps/arweave/lib/RandomX/build512 ]; then make -C apps/arweave/lib/RandomX/build512 clean; fi\""},
+	{"(freebsd|netbsd|openbsd)", clean, "bash -c \"if [ -d apps/arweave/lib/RandomX/build512 ]; then gmake -C apps/arweave/lib/RandomX/build512 clean; fi\""},
+	% Clean randomx4096
+	{"(linux|darwin)", clean, "bash -c \"if [ -d apps/arweave/lib/RandomX/build4096 ]; then make -C apps/arweave/lib/RandomX/build4096 clean; fi\""},
+	{"(freebsd|netbsd|openbsd)", clean, "bash -c \"if [ -d apps/arweave/lib/RandomX/build4096 ]; then gmake -C apps/arweave/lib/RandomX/build4096 clean; fi\""},
+	% Clan NIFs
 	{"(linux|darwin)", clean, "make -C apps/arweave/c_src clean"},
 	{"(freebsd|netbsd|openbsd)", clean, "gmake -C apps/arweave/c_src clean"}
 ]}.
